Stat module cllean up.
