+++ src/modules/ui/ProjectListRows.js
@@ -181,7 +181,7 @@ const TitleLinkLearnMore = styled.a`
 `
 
 const IcoArrow = styled.span`
-display: inline-block
+display: inline-block;
 transition: 0.15s ease-in-out;
 color: var(--highlight-color--makerteal);
 `
